Refactor fetch/cache from upstream into eth-loader class.
